Move JS rules to their own folder
