+++ src/rules/javascript/index.js
