+++ src/rules/javascript/mozIndexedDB.js
