+++ src/rules/javascript/mozIndexedDB_property.js
